# openapi.yaml
openapi: 3.1.0
info:
  title: Sender API
  version: 0.1.0
  description: Multi-tenant email API backed by Amazon SES.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /emails:
    get:
      summary: List emails
      parameters:
        - $ref: '#/components/parameters/TeamID'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200': { $ref: '#/components/responses/EmailList' }
    post:
      summary: Queue an email
      parameters:
        - $ref: '#/components/parameters/TeamID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { $ref: '#/components/requestBodies/SendEmail' }
      responses:
        '201': { $ref: '#/components/responses/EmailCreated' }
        '200': { $ref: '#/components/responses/EmailCreated' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /emails/batch:
    post:
      summary: Queue multiple emails
      parameters:
        - $ref: '#/components/parameters/TeamID'
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/SendEmailRequest' }
      responses:
        '201': { $ref: '#/components/responses/JSON' }
        '207': { $ref: '#/components/responses/JSON' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /emails/dead-letters:
    get:
      summary: List failed queue items for the current team
      parameters:
        - $ref: '#/components/parameters/TeamID'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/JSON' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /emails/dead-letters/{id}/replay:
    parameters:
      - $ref: '#/components/parameters/ID'
      - $ref: '#/components/parameters/TeamID'
    post:
      summary: Replay a failed queue item
      responses:
        '202': { $ref: '#/components/responses/JSON' }
        '400': { $ref: '#/components/responses/Error' }
  /emails/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
      - $ref: '#/components/parameters/TeamID'
    get:
      summary: Get an email
      responses: { '200': { $ref: '#/components/responses/JSON' } }
    delete:
      summary: Cancel a queued email
      responses: { '204': { description: Cancelled } }
  /emails/{id}/events:
    parameters:
      - $ref: '#/components/parameters/ID'
      - $ref: '#/components/parameters/TeamID'
    get:
      summary: List email events
      responses: { '200': { $ref: '#/components/responses/JSON' } }
...
  /contacts:
    get:
      summary: List contacts
      parameters: [ { $ref: '#/components/parameters/TeamID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Offset' } ]
      responses: { '200': { $ref: '#/components/responses/JSON' } }
    post:
      summary: Create a contact
      parameters: [ { $ref: '#/components/parameters/TeamID' } ]
      requestBody: { $ref: '#/components/requestBodies/JSON' }
      responses: { '201': { $ref: '#/components/responses/JSON' } }
  /contacts/import:
    post:
      summary: Import contacts from CSV
      parameters: [ { $ref: '#/components/parameters/TeamID' } ]
      requestBody:
        required: true
        content:
          multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } }
      responses: { '201': { $ref: '#/components/responses/JSON' } }
  /contacts/{id}:
    parameters: [ { $ref: '#/components/parameters/ID' }, { $ref: '#/components/parameters/TeamID' } ]
    get:
      summary: Get a contact
      responses: { '200': { $ref: '#/components/responses/JSON' } }
    patch:
      summary: Update a contact
      requestBody: { $ref: '#/components/requestBodies/JSON' }
      responses: { '200': { $ref: '#/components/responses/JSON' } }
    delete:
      summary: Delete a contact
      responses: { '204': { description: Deleted } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or re_ API key
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    TeamID: { name: X-Team-ID, in: header, required: false, description: Required for JWT requests; API keys already identify their team., schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 255 } }
    RequiredIdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    Offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
  requestBodies:
    JSON:
      required: true
      content: { application/json: { schema: { type: object } } }
    SendEmail:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/SendEmailRequest' } } }
  responses:
    JSON: { description: JSON response, content: { application/json: { schema: {} } } }
    Error: { description: Error response, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    EmailList: { description: Email list, content: { application/json: { schema: { type: object } } } }
    EmailCreated: { description: Queued email, content: { application/json: { schema: { $ref: '#/components/schemas/EmailResponse' } } } }
    Conflict: { description: Idempotency conflict, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited: { description: Daily recipient or request rate limit exceeded, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unavailable: { description: Queue or usage limiter unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    SendEmailRequest:
      type: object
      required: [from, to, subject]
      properties:
        from: { type: string, format: email }
        to: { type: array, minItems: 1, maxItems: 50, items: { type: string, format: email } }
        cc: { type: array, items: { type: string, format: email } }
        bcc: { type: array, items: { type: string, format: email } }
        subject: { type: string, maxLength: 998 }
        category: { type: string, enum: [transactional, marketing], default: transactional, description: Marketing messages require one recipient, verified DMARC, and automatically receive RFC 2369/8058 unsubscribe headers. }
        text: { type: string }
        html: { type: string }
        reply_to: { type: array, items: { type: string, format: email } }
        headers: { type: object, additionalProperties: { type: string } }
        metadata: { type: object, additionalProperties: { type: string } }
        tags: { type: array, items: { type: object } }
        attachments: { type: array, items: { type: object } }
        scheduled_at: { type: string, format: date-time }
...
    EmailResponse:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
        idempotent: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }